refactor: Move test files
