+++ src/test/java/racingcar/domain/RaceTest.java
@@ -1,17 +1,18 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+import java.util.Collections;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Race;
 
 class RaceTest {
 
     @Test
     void _0에서_9사이의_정수를_반환하는지_확인한다() {
         int[] check = new int[10];
+        Race race = new Race(Collections.singletonList("car"), 7);
         for (int i = 0; i < 100; i++) {
-            int randomNumber = Race.getRandomNumber();
-            System.out.println(randomNumber);
+            int randomNumber = race.getRandomNumber();
             check[randomNumber]++;
         }
         int sum = 0;
